+++ serlib/ser_retroknowledge.mli
@@ -12,8 +12,9 @@
 (************************************************************************)
 (* Status: Experimental                                                 *)
 (************************************************************************)
+open Sexplib
 
-type retroknowledge = [%import: Retroknowledge.retroknowledge]
+type retroknowledge = Retroknowledge.retroknowledge
 
-let sexp_of_retroknowledge = Serlib_base.sexp_of_opaque ~typ:"Retroknowledge.retroknowledge"
-let retroknowledge_of_sexp = Serlib_base.opaque_of_sexp ~typ:"Retroknowledge.retroknowledge"
+val sexp_of_retroknowledge : retroknowledge -> Sexp.t
+val retroknowledge_of_sexp : Sexp.t -> retroknowledge